Fix bug in scrolling of left panel
Scrolling of the left panel (table of contents) was broken in the
rare cases where the height of left panel is much bigger than the
height of the actual page contents. This can happen in Reaktor
Library because the "Pages in folder X" listing has been moved
from the bottom of the content section to the left panel, below
table of contents.

By default, the `main.js` script attaches a `d-fixed` class on the
left panel to make the table of contents hover next to the
contents as the page is scrolled down. However, if the left panel
height is much bigger than the height of the contents (for example
when there are a lot of files in a folder, but there is no text
content on the index page itself), this `d-fixed` class actually
breaks the scrolling and the list cannot be scrolled all the way
to bottom.

The fix is to simply compare the heights of the left panel and
content elements, and only attach the `d-fixed` class when content
element's height is bigger than left panel element height.