Fix bugs in code block formatting
There was a bug in code block formatting. When code block did not
have language defined (on the first line of the code block, after
the three back ticks), the second line of the code block (the
first actual line of code) was cut off.

Now the first line of code is preserved even when there is no
language defined for the code block.

Another problem was that whitespace before language definition
caused the formatting to not remove the language definition from
the code block contents. Adding one `.trim()` seemed to do the
trick.