Remove wacky sorting functionality
Sorting of documents and folders was working in an unexpected and
unintuitive way: a number _anywhere_ in the doc/folder name was
being used as the sorting term, instead of the title. For example,
"Two factor authentication (2FA)" was sorted before "Admin FAQ"
because of the "2" in the name of the former.

Now this odd sorting mechanism has been removed and the documents
and folders are simply sorted alphabetically by their name.